+++ deepr/metrics/core.py
@@ -1,5 +1,6 @@
 """Last value metric"""
 
+import logging
 from typing import Dict, List, Tuple
 
 import tensorflow as tf
@@ -7,37 +8,38 @@ import tensorflow as tf
 from deepr.metrics import base
 
 
+LOGGER = logging.getLogger(__name__)
+
+
 class LastValue(base.Metric):
     """Last value Metric"""
 
-    def __init__(self, tensors: List[str] = None):
+    def __init__(self, tensors: List[str] = None, pattern: str = None):
         self.tensors = tensors
+        self.pattern = pattern
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}(tensors={self.tensors}, pattern={self.pattern})"
 
     def __call__(self, tensors: Dict[str, tf.Tensor]) -> Dict[str, Tuple]:
-        if self.tensors is None:
-            tensors = {key: tensor for key, tensor in tensors.items() if len(tensor.shape) == 0}
-        else:
-            tensors = {name: tensors[name] for name in self.tensors}
-        return {name: last_value_metric(value, name) for name, value in tensors.items()}
-
-
-def last_value_metric(value, name):
-    last_value = base.get_metric_variable(name=f"{name}_last", shape=(), dtype=value.dtype)
-    update_op = tf.assign(last_value, value)
-    return (last_value, update_op)
+        tensors = base.get_scalars(tensors, names=self.tensors, pattern=self.pattern)
+        LOGGER.info(f"{self} -> {', '.join(tensors.keys())}")
+        return {name: (tensor, tf.no_op()) for name, tensor in tensors.items()}
 
 
 class MaxValue(base.Metric):
     """Max value Metric"""
 
-    def __init__(self, tensors: List[str] = None):
+    def __init__(self, tensors: List[str] = None, pattern: str = None):
         self.tensors = tensors
+        self.pattern = pattern
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}(tensors={self.tensors}, pattern={self.pattern})"
 
     def __call__(self, tensors: Dict[str, tf.Tensor]) -> Dict[str, Tuple]:
-        if self.tensors is None:
-            tensors = {key: tensor for key, tensor in tensors.items() if len(tensor.shape) == 0}
-        else:
-            tensors = {name: tensors[name] for name in self.tensors}
+        tensors = base.get_scalars(tensors, names=self.tensors, pattern=self.pattern)
+        LOGGER.info(f"{self} -> {', '.join(tensors.keys())}")
         return {name: max_value_metric(value, name) for name, value in tensors.items()}