+++ deepr/metrics/step.py
@@ -13,6 +13,9 @@ class StepCounter(base.Metric):
     def __init__(self, name: str):
         self.name = name
 
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}(name={self.name})"
+
     def __call__(self, tensors: Dict[str, tf.Tensor]) -> Dict[str, Tuple]:
         # pylint: disable=unused-argument
         value = base.get_metric_variable(name=self.name, shape=(), dtype=tf.int64)